+++ django/contrib/contenttypes/views.py
@@ -2,7 +2,8 @@ from __future__ import unicode_literals
 
 from django import http
 from django.contrib.contenttypes.models import ContentType
-from django.contrib.sites.models import Site, get_current_site
+from django.contrib.sites.models import Site
+from django.contrib.sites.shortcuts import get_current_site
 from django.core.exceptions import ObjectDoesNotExist
 from django.utils.translation import ugettext as _
